task_health_monitor.sh: add OFW specific support for restarting lighttpd ( OFW-6067 )
[Issue]
The lighttpd process crashed and Selfheal was trying to restart it by running the script webgui.sh. webgui.sh is not present in the device and lighttpd is never restarted.

[Solution]
Make Selfheal to restart lighttpd by running the script start_lighttpd.sh in restart mode.

[Testing]
1. DUT is up and running and GUI is accessible
2. Kill lighttpd process through command line
3. Check lighttpd process after 10 minutes and it should be running with a new PID
4. GUI should be accessible.

Upstream Candidate: No. LG Specific feature.

+++ scripts/task_health_monitor.sh
@@ -2514,8 +2514,12 @@ case $SELFHEAL_TYPE in
                 echo_t "RDKB_PROCESS_CRASHED : lighttpd is not running, restarting it"
                 t2CountNotify "SYS_SH_lighttpdCrash"
                 #lighttpd -f $LIGHTTPD_CONF
-                rm /tmp/webgui.lock
-                sh /etc/webgui.sh
+                if [ "$FIRMWARE_TYPE" = "OFW" ]; then
+                    sh /etc/start_lighttpd.sh restart
+                else
+                    rm /tmp/webgui.lock
+                    sh /etc/webgui.sh
+                fi
             fi
         fi
     ;;
